Move unit tests into standalone files
